refactor: split MCP dev server + upgrade to registerTool/registerPrompt API
- Split server.ts (800+ lines) into 7 focused modules: discovery-tools,
  validation-tools, utility-tools, resources, prompts, helpers, schema
- Migrate all deprecated server.tool() to server.registerTool()
- Migrate all deprecated server.prompt() to server.registerPrompt()
- Fix zod v4/v3 type compatibility via zod/v3 compat layer (SDK issue #796)
- Remove all `as any` casts from MCP tool/prompt schemas
- Upgrade @modelcontextprotocol/sdk to 1.29.0
- Add quality gate rules: 500-line file limit, context7 API verification
- Add file size rule to architecture skill
