+++ addons/adapter-mcp/cap-adapter-mcp/package.json
@@ -13,7 +13,7 @@
     "test": "bun test"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.27.1"
+    "@modelcontextprotocol/sdk": "^1.29.0"
   },
   "peerDependencies": {
     "@linchkit/core": "^0.1.0"
