+++ package.json
@@ -35,6 +35,7 @@
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
+    "@modelcontextprotocol/sdk": "1.29.0",
     "@restatedev/restate-sdk": "1.11.1",
     "@tanstack/react-table": "^8.21.3",
     "ai": "^6.0.134",
