+++ packages/cli/src/templates/index.ts
@@ -765,6 +765,18 @@ bun test
 \`\`\`
 All tests must pass. Write tests alongside implementation.
 
+## 5. File Size Check
+Single files MUST NOT exceed 500 lines. If a file is too large:
+- Split by responsibility (e.g. \`discovery-tools.ts\`, \`validation-tools.ts\`)
+- Extract shared helpers to a separate file
+- Each file should have one clear responsibility
+
+## Dependency API Verification
+When using third-party library APIs, MUST verify current usage via context7 MCP tools:
+1. \`resolve-library-id\` — find the library's context7 ID
+2. \`query-docs\` — query the specific API pattern
+Do NOT rely on training data — libraries change frequently.
+
 ## Commit Convention
 Use Conventional Commits: \`feat:\`, \`fix:\`, \`refactor:\`, \`test:\`, \`docs:\`, \`chore:\`
 `,
@@ -846,6 +858,12 @@ Before adding to core, ask: "Without this, is a zero-capability LinchKit still A
 - \`ui\` MUST NOT import from \`server\`
 - No circular dependencies
 - Dependency flows one way: Capability → Core
+
+## File Size Rule
+- Single file MUST NOT exceed 500 lines
+- Split by responsibility when approaching the limit
+- Shared helpers go to separate files
+- \`index.ts\` only re-exports, no implementation logic
 `,
     },
   ];
